refactor: solve circule dependency
