+++ src/TreeNode.tsx
@@ -3,7 +3,7 @@ import classNames from 'classnames';
 import pickAttrs from 'rc-util/lib/pickAttrs';
 // @ts-ignore
 import { TreeContext, TreeContextProps } from './contextTypes';
-import { IconType, Key, DataNode, BasicDataNode } from './interface';
+import { TreeNodeProps } from './interface';
 import Indent from './Indent';
 import { convertNodePropsToEventData } from './utils/treeUtil';
 
@@ -11,44 +11,6 @@ const ICON_OPEN = 'open';
 const ICON_CLOSE = 'close';
 
 const defaultTitle = '---';
-
-export interface TreeNodeProps<TreeDataType extends BasicDataNode = DataNode> {
-  eventKey?: Key; // Pass by parent `cloneElement`
-  prefixCls?: string;
-  className?: string;
-  style?: React.CSSProperties;
-  id?: string;
-
-  // By parent
-  expanded?: boolean;
-  selected?: boolean;
-  checked?: boolean;
-  loaded?: boolean;
-  loading?: boolean;
-  halfChecked?: boolean;
-  title?: React.ReactNode | ((data: TreeDataType) => React.ReactNode);
-  dragOver?: boolean;
-  dragOverGapTop?: boolean;
-  dragOverGapBottom?: boolean;
-  pos?: string;
-  domRef?: React.Ref<HTMLDivElement>;
-  /** New added in Tree for easy data access */
-  data?: TreeDataType;
-  isStart?: boolean[];
-  isEnd?: boolean[];
-  active?: boolean;
-  onMouseMove?: React.MouseEventHandler<HTMLDivElement>;
-
-  // By user
-  isLeaf?: boolean;
-  checkable?: boolean;
-  selectable?: boolean;
-  disabled?: boolean;
-  disableCheckbox?: boolean;
-  icon?: IconType;
-  switcherIcon?: IconType;
-  children?: React.ReactNode;
-}
 
 export interface InternalTreeNodeProps extends TreeNodeProps {
   context?: TreeContextProps;
