+++ src/index.ts
@@ -1,8 +1,7 @@
 import Tree from './Tree';
 import TreeNode from './TreeNode';
 import type { TreeProps } from './Tree';
-import type { TreeNodeProps } from './TreeNode';
-import type { BasicDataNode, FieldDataNode } from './interface';
+import type { TreeNodeProps, BasicDataNode, FieldDataNode } from './interface';
 
 export { TreeNode };
 export type { TreeProps, TreeNodeProps, BasicDataNode, FieldDataNode };
