+++ src/interface.tsx
@@ -1,7 +1,43 @@
 import * as React from 'react';
-import { TreeNodeProps } from './TreeNode';
 
 export { ScrollTo } from 'rc-virtual-list/lib/List';
+export interface TreeNodeProps<TreeDataType extends BasicDataNode = DataNode> {
+  eventKey?: Key; // Pass by parent `cloneElement`
+  prefixCls?: string;
+  className?: string;
+  style?: React.CSSProperties;
+  id?: string;
+
+  // By parent
+  expanded?: boolean;
+  selected?: boolean;
+  checked?: boolean;
+  loaded?: boolean;
+  loading?: boolean;
+  halfChecked?: boolean;
+  title?: React.ReactNode | ((data: TreeDataType) => React.ReactNode);
+  dragOver?: boolean;
+  dragOverGapTop?: boolean;
+  dragOverGapBottom?: boolean;
+  pos?: string;
+  domRef?: React.Ref<HTMLDivElement>;
+  /** New added in Tree for easy data access */
+  data?: TreeDataType;
+  isStart?: boolean[];
+  isEnd?: boolean[];
+  active?: boolean;
+  onMouseMove?: React.MouseEventHandler<HTMLDivElement>;
+
+  // By user
+  isLeaf?: boolean;
+  checkable?: boolean;
+  selectable?: boolean;
+  disabled?: boolean;
+  disableCheckbox?: boolean;
+  icon?: IconType;
+  switcherIcon?: IconType;
+  children?: React.ReactNode;
+}
 
 /** For fieldNames, we provides a abstract interface */
 export interface BasicDataNode {
